pyproject.toml
[project]
name = "coa-consumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.7",
    "python-dotenv>=1.1.1",
    "pyrad>=2.4",
]
